Removed the first bugs from the recursive aggregates. They now work for at least one test example.
I again stumbled upon the problem of very slow resolution of blockednqueens, so for documentation purposes: blockednqueens tends to have bad performance (compared to older versions of minisatid) because of the way exists unique statements are handled. When extra watches are added (see comments in code)  and clause reordering is done upon EU propagation, some instances of the problem are solved much faster. But there are also instances where it performs worse, so the choice was made to mainly use the clearest code (without extra watches). In the future it might be useful to look at it again and derive a good heuristic on when to use which code.