Refactored constraint solver into a managable framework a la aggsolver and added further support for ao binary arithmetic constraints (also some bugfixing there).
Separated unittests from the rest of the code.

Added a reader for ground lparse format (based on the smodels source code), which is fully implemented except for optimization statements (only basically tested).

Streamlined idpexception use throughout the program and in the parser, so that the printed error messages are uniform and parsing info is printed also whenever relevant.

Restructured the parser a bit more, using enums when possible instead of cast ints.