INCORRECT. Still some failing benches to check out! Changes to PBSolver: 	replaced long long with int64_t and analogous for unsigned version. 	took care of [0] initialized arrays and effectively missing includes. 	Could NOT fix arrays initialized with variables, which is also not allowed in iso C++ -> this is only remaining issue with iso c++ errors.
Rewritten argument parsing using tclap and in new files. Not really code reduction, but certainly more readable and maintainable. Putting the arguments in the separate classes was not really usefull as the global option class which is passed around needs to have knowledge of them anyway and has to be initialized explicitly in main. Renamed ECNF_mode to SolverOption.

Moved checking for well-foundedness of definition to definition solver, adding generic checkStatus above it.