Further correcting lazy grounding
When doing model expansion, force adaptive search strategy.
	This allows to delay finishparsing until really looking for loops, which might decrease time spend in that part of the algorithm.
Finishparsing in eventqueue now uses a notification system, such that iterative grounding and finishing does not pose problems.
When a total model is found, missing id propagation might lead to grounding and backtracking, so should take this into account in both the defsolver and the eventqueue.