Refactored visitors
Added all externally usable visitors to external and refactored the structure accordingly.
There are still some issues with the template Stream, which now has to be instantiated a compile time, but that is probably inevitable.
Some parser changes.
Fully implemented the ECNF printer except for the translation and made it also externally accessible.