Lots of fixes
Fixed calculation of number of formulas for CP constraints.
Removed intvars from the eternel propagators, strongly reducing the cost of updating bounds and thus of any computation with fd vars.
Fixed resetting the state when root unit literals are still present, which was existence-checked on the root unit literal element instead of the literal itself (as the element can change during its lifetime).
Fixed the timetrail, which could not do a noop "backtrack" to the current level.
Added a new higher priority to the eventqueue, to which fd vars have been added.
Removed an unused method from searchengine.
Fixed a bug where during clause relocation, their references in root unit lits were not updated, later on crashing the solver when they were accessed.