Resolved long standing issues with pbsolver
Removed spurious code.
Moved all global fields to a PBOptions class, which is kept by PbSolver. Updated the usage of the changed interface.
Commented the clause splitting code (BUGGED!).
Moved the static field in the clausifier to a non-static position in the Solver (enabling it to keep just enough state).