Added internal id support, bugfix in idsolver, renames
Renamed Var to Atom
Renamed Literal to Lit, removing the distinction between internal and external Literals, which made it difficult to understand
Moved an assert to a justification not being empty to the correct location (it can be empty for aggregates).
Generalized the Remapper to generate uint identifiers and maintain a mapping to an original one wherever relevant. Added support to generate internal variable numbers too, by remapping all variables in extAdd.
Added support in the LparseTranslator for finite domain variable translation.