Fixes bug in product propagator when negative values are present
The problem with negative values was in the propagateWithoutNeg method
That method was written in the assumption that in ANY case, those variables
had to be positive. Therefore, clauses learned by this method were only valid
in parts of the search space where all variables are postitive.
I adapted all its learned clauses to include this information.

This fix is needed (among others) for the 11SATcpprodEQ-test.
HOWEVER, this test will not work yet, since I also found another bug ;-)